Add SubjectViewerStore to task stories Move the `createStore` function into a shared helper. Add the SubjectViewerStore to the mock store, so that we can mock the subject loading state.
