+++ packages/lib-classifier/src/plugins/tasks/helpers/createStore/index.js
@@ -1,0 +1,1 @@
+export { default } from './createStore'
